Refactoring code into appropriate layers/files
