+++ src/repository/stockRepository.js
@@ -1,22 +1,21 @@
 var sql = require("mssql");
+var { connectionString } = require('../config/config.js');
 
-export default class StockRepository {
+exports.getAllStockUsers = function() {
+    return new Promise((resolve, reject) => { sql.connect(connectionString).then(function() {
+        var request = new sql.Request();
 
-    async GetStocksWithUsers() {
-
-        var config = {
-            user: '',
-            password: '',
-            server: '',
-            database: ''
-        }
-
-        sql.connect(config, function(err) {
-            var request = sql.Request();
-
-            request.query('select * from stocks', function(error, recordset) {
-                return recordset;
-            });
+        request.query(`SELECT sm.Ticker, sm.CompanyName, u.FirstName, u.LastName, u.EmailAddress
+                       FROM StockMetadata sm (NOLOCK)
+                       INNER JOIN UserStock us (NOLOCK)
+                       ON us.StockID = sm.StockID
+                       INNER JOIN [User] u (NOLOCK)
+                       ON u.UserId = us.UserID`, function(error, recordset) {
+            
+        var stocksFromDb = recordset.recordsets[0];
+        resolve(stocksFromDb);
+            
         });
-    }
+    })
+});
 }
